+++ package/recalbox-romfs2/scripts/systemholder.py
@@ -25,7 +25,7 @@ class SystemHolder:
             "mandatory": 3,
         }
 
-        def __init__(self, systemtype: str, pad: str, keyboard: str, mouse: str, lightgun: str, releasedate: str, retroachievements: bool, crtmultiresolution: bool, crtmultiregion: bool):
+        def __init__(self, systemtype: str, pad: str, keyboard: str, mouse: str, lightgun: str, releasedate: str, manufacturer : str, retroachievements: bool, crtmultiresolution: bool, crtmultiregion: bool):
             if systemtype not in self.__systemTypes.keys(): raise TypeError("Invalid system type! {}".format(systemtype))
             if pad not in self.__deviceRequirement.keys(): raise TypeError("Invalid pad type! {}".format(pad))
             if keyboard not in self.__deviceRequirement.keys(): raise TypeError("Invalid keyboard type! {}".format(keyboard))
@@ -37,6 +37,7 @@ class SystemHolder:
             self.__mouse: str = mouse
             self.__lightgun: str = lightgun
             self.__releasedate: str = releasedate
+            self.__manufacturer = manufacturer
             self.__retroachievements: bool = retroachievements
             self.__multiresolution: bool = crtmultiresolution
             self.__multiregion: bool = crtmultiregion
@@ -49,6 +50,7 @@ class SystemHolder:
                 "mouse": self.__mouse,
                 "lightgun": self.__lightgun,
                 "releasedate": self.__releasedate,
+                "manufacturer": self.__manufacturer,
                 "retroachievements": '1' if self.__retroachievements else '0',
                 "crt.multiresolution": '1' if self.__multiresolution else '0',
                 "crt.multiregion": '1' if self.__multiregion else '0'
@@ -86,6 +88,9 @@ class SystemHolder:
 
         @property
         def ReleaseDate(self): return self.__releasedate
+
+        @property
+        def Manufacturer(self): return self.__manufacturer
 
         @property
         def RetroAchievements(self) -> bool: return self.__retroachievements
@@ -183,7 +188,7 @@ class SystemHolder:
         }
         self.__port: bool = False
         self.__readOnly: bool = False
-        self.__properties = SystemHolder.SystemProperties("virtual", "no", "no", "no", "no", "", False, False, False)
+        self.__properties = SystemHolder.SystemProperties("virtual", "no", "no", "no", "no", "", "virtual", False, False, False)
         self.__coreLists: Dict[str, List[SystemHolder.Core]] = dict()
         self.__coreCount: int = 0
         self.__deserialize()
@@ -324,6 +329,7 @@ class SystemHolder:
             mouse=self.__get(desc, "properties", "device.mouse", "no", True),
             lightgun=self.__get(desc, "properties", "device.lightgun", "no", True),
             releasedate=self.__get(desc, "properties", "release.date", "", True),
+            manufacturer=self.__get(desc, "properties", "manufacturer", "", True),
             retroachievements=(self.__get(desc, "properties", "retroachievements", "0", True) == '1'),
             crtmultiresolution=(self.__get(desc, "properties", "crt.multiresolution", "0", True) == '1'),
             crtmultiregion=(self.__get(desc, "properties", "crt.multiregion", "0", True) == '1'),