+++ projects/frontend/es-app/src/guis/MenuMessages.h
@@ -112,6 +112,7 @@ class MenuMessages
     #define MENUMESSAGE_ADVANCED_EMULATOR_ADVANCED_HELP_MSG "Override global options like emulator, core, ratio and more for each available system in your recalbox."
     #define MENUMESSAGE_ADVANCED_BOOT_HELP_MSG "Configure boot options that make your recalbox boot straight into a system or into Kodi, lock a user to a single system, or directly show the gamelist."
     #define MENUMESSAGE_ADVANCED_ALLGAMES_HELP_MSG "Show a 'all-games' system with all games from all systems."
+    #define MENUMESSAGE_ADVANCED_SORTINGOPTION_HELP_MSG "Default: use original or custom systemlist.xml order\nSystem Type: order by Console/Handheld/Computer/Arcade/Other\nRelease Date: order by release date asc\nManufacturer: order by manufacturer (e.g. Sega)\nSpecial Blend: Sort by type then Manufacturer then Release Date"
     #define MENUMESSAGE_ADVANCED_MULTIPLAYERS_HELP_MSG "Show multiplayer games (all games playable by two or more players)."
     #define MENUMESSAGE_ADVANCED_LASTPLAYED_HELP_MSG "Show last played games."
     #define MENUMESSAGE_ADVANCED_LIGHTGUN_HELP_MSG "Show all available games playable with a lightgun."