+++ projects/frontend/es-app/src/guis/menus/GuiMenuUserInterface.cpp
@@ -14,25 +14,14 @@
 GuiMenuUserInterface::GuiMenuUserInterface(WindowManager& window, SystemManager& systemManager)
   : GuiMenuBase(window, _("UI SETTINGS"), this)
   , mSystemManager(systemManager)
+  , mOriginalSort(RecalboxConf::Instance().GetSystemSorting())
 {
-  // Brightness
-  if (Board::Instance().HasBrightnessSupport())
-    mBrightness = AddSlider(_("BRIGHTNESS"), 0.f, 8.f, 1.f, (float)RecalboxConf::Instance().GetBrightness(), "", (int)Components::Brightness, this);
-
   // Screensavers
   AddSubMenu(_("SCREENSAVER"), (int)Components::ScreenSaver, _(MENUMESSAGE_UI_SCREENSAVER_HELP_MSG));
 
-  // display clock
-  mClock = AddSwitch(_("CLOCK IN MENU"), RecalboxConf::Instance().GetClock(), (int)Components::Clock, this, _(MENUMESSAGE_UI_CLOCK_HELP_MSG));
-
-  // show help
-  mHelp = AddSwitch(_("ON-SCREEN HELP"), RecalboxConf::Instance().GetShowHelp(), (int)Components::Help, this, _(MENUMESSAGE_UI_ONSCREENHELP_HELP_MSG));
-
-  // Popup settings
-  AddSubMenu(_("POPUP SETTINGS"), (int)Components::Popups, _(MENUMESSAGE_UI_POPUP_HELP_MSG));
-
-  // quick system select (left/right in game list view)
-  mQuickSelect = AddSwitch(_("QUICK SYSTEM SELECT"), RecalboxConf::Instance().GetQuickSystemSelect(), (int)Components::QuickSelect, this, _(MENUMESSAGE_UI_QUICK_HELP_MSG));
+  // Brightness
+  if (Board::Instance().HasBrightnessSupport())
+      mBrightness = AddSlider(_("BRIGHTNESS"), 0.f, 8.f, 1.f, (float)RecalboxConf::Instance().GetBrightness(), "", (int)Components::Brightness, this);
 
   // Theme
   AddSubMenu(_("THEME"), (int)Components::Theme, _(MENUMESSAGE_UI_THEME_HELP_MSG));
@@ -40,11 +29,26 @@ GuiMenuUserInterface::GuiMenuUserInterface(WindowManager& window, SystemManager&
   // Theme Options
   AddSubMenu(_("THEME CONFIGURATION"), (int)Components::ThemeConfig, _(MENUMESSAGE_UI_THEME_CONFIGURATION_MSG));
 
-  // Game List Update
-  AddSubMenu(_("UPDATE GAMES LISTS"), (int)Components::UpdateGamelist, _(MENUMESSAGE_UI_UPDATE_GAMELIST_HELP_MSG));
+  // System sort
+  mSort = AddList<SystemSorting>(_("SYSTEM SORTING"), (int)Components::SystemSort, this, GetSortingEntries(), _(MENUMESSAGE_ADVANCED_SORTINGOPTION_HELP_MSG));
+
+  // quick system select (left/right in game list view)
+  mQuickSelect = AddSwitch(_("QUICK SYSTEM SELECT"), RecalboxConf::Instance().GetQuickSystemSelect(), (int)Components::QuickSelect, this, _(MENUMESSAGE_UI_QUICK_HELP_MSG));
+
+  // show help
+  mHelp = AddSwitch(_("ON-SCREEN HELP"), RecalboxConf::Instance().GetShowHelp(), (int)Components::Help, this, _(MENUMESSAGE_UI_ONSCREENHELP_HELP_MSG));
 
   // Swap validate and cancel buttons
   mSwapValidateAndCancel = AddSwitch(_("SWAP VALIDATE/CANCEL BUTTONS"), RecalboxConf::Instance().GetSwapValidateAndCancel(), (int)Components::SwapValidateAndCancel, this, _(MENUMESSAGE_UI_SWAP_VALIDATE_CANCEL_BUTTONS_HELP_MSG));
+
+  // display clock
+  mClock = AddSwitch(_("CLOCK IN MENU"), RecalboxConf::Instance().GetClock(), (int)Components::Clock, this, _(MENUMESSAGE_UI_CLOCK_HELP_MSG));
+
+  // Popup settings
+  AddSubMenu(_("POPUP SETTINGS"), (int)Components::Popups, _(MENUMESSAGE_UI_POPUP_HELP_MSG));
+
+  // Game List Update
+  AddSubMenu(_("UPDATE GAMES LISTS"), (int)Components::UpdateGamelist, _(MENUMESSAGE_UI_UPDATE_GAMELIST_HELP_MSG));
 }
 
 void GuiMenuUserInterface::ReloadGamelists()
@@ -80,6 +84,7 @@ void GuiMenuUserInterface::SubMenuSelected(int id)
     case Components::Clock:
     case Components::SwapValidateAndCancel:
     case Components::Help:
+    case Components::SystemSort:
     case Components::QuickSelect: break;
   }
 }
@@ -112,6 +117,35 @@ void GuiMenuUserInterface::SwitchComponentChanged(int id, bool status)
     case Components::ThemeConfig:
     case Components::UpdateGamelist:
     case Components::ScreenSaver:
+    case Components::SystemSort:
     case Components::Brightness: break;
   }
+}
+
+
+void GuiMenuUserInterface::OptionListComponentChanged(int id, int index, const SystemSorting& value)
+{
+    (void)index;
+    (void)id;
+    RecalboxConf::Instance().SetSystemSorting(value).Save();
+
+    mSystemManager.SystemSorting();
+    ViewController::Instance().getSystemListView().Sort();
+}
+
+std::vector<GuiMenuBase::ListEntry<SystemSorting>> GuiMenuUserInterface::GetSortingEntries()
+{
+    mOriginalSort = RecalboxConf::Instance().GetSystemSorting();
+    return std::vector<ListEntry<SystemSorting>>
+    ({
+        { _("DEFAULT")                                    , SystemSorting::Default                                    , mOriginalSort == SystemSorting::Default },
+        { _("NAME")                                       , SystemSorting::Name                                       , mOriginalSort == SystemSorting::Name },
+        { _("RELEASE DATE")                               , SystemSorting::ReleaseDate                                , mOriginalSort == SystemSorting::ReleaseDate },
+        { _("TYPE, THEN NAME")                            , SystemSorting::SystemTypeThenName                         , mOriginalSort == SystemSorting::SystemTypeThenName },
+        { _("TYPE, THEN RELEASE DATE")                    , SystemSorting::SystemTypeThenReleaseDate                  , mOriginalSort == SystemSorting::SystemTypeThenReleaseDate },
+        { _("MANUFACTURER, THEN NAME")                    , SystemSorting::ManufacturerThenName                       , mOriginalSort == SystemSorting::ManufacturerThenName },
+        { _("MANUFACTURER, THEN RELEASE DATE")            , SystemSorting::ManufacturerThenReleaseData                , mOriginalSort == SystemSorting::ManufacturerThenReleaseData },
+        { _("TYPE, THEN MANUFACTURER, THEN NAME")         , SystemSorting::SystemTypeThenManufacturerThenName         , mOriginalSort == SystemSorting::SystemTypeThenManufacturerThenName },
+        { _("TYPE, THEN MANUFACTURER, THEN RELEASE DATE") , SystemSorting::SystemTypeThenManufacturerThenReleasdeDate , mOriginalSort == SystemSorting::SystemTypeThenManufacturerThenReleasdeDate },
+    });
 }