+++ projects/frontend/es-core/src/RecalboxConf.h
@@ -10,6 +10,7 @@
 #include <scraping/scrapers/screenscraper/ScreenScraperEnums.h>
 #include <scraping/scrapers/screenscraper/Languages.h>
 #include <audio/AudioMode.h>
+#include "systems/SystemSorting.h"
 
 // Forward declaration
 class SystemData;
@@ -108,6 +109,7 @@ class RecalboxConf: public IniFile, public StaticLifeCycleControler<RecalboxConf
     DefineGetterSetterEnum(ScreenScraperThumbnail, ScreenScraperEnums::ScreenScraperImageType, sScreenScraperThumbnail, ScreenScraperEnums::ScreenScraperImageType)
     DefineGetterSetterEnum(ScreenScraperVideo, ScreenScraperEnums::ScreenScraperVideoType, sScreenScraperVideo, ScreenScraperEnums::ScreenScraperVideoType)
     DefineGetterSetterEnum(AudioMode, AudioMode, sAudioOptions, AudioModeTools::AudioMode)
+    DefineGetterSetterEnum(SystemSorting, SystemSorting, sSystemSorting, SystemSorting)
 
     DefineGetterSetter(Hostname, std::string, String, sHostname, "RECALBOX")
 
@@ -349,6 +351,8 @@ class RecalboxConf: public IniFile, public StaticLifeCycleControler<RecalboxConf
     static constexpr const char* sFavoritesOnly              = "emulationstation.favoritesonly";
     static constexpr const char* sShowHidden                 = "emulationstation.showhidden";
 
+    static constexpr const char* sSystemSorting              = "emulationstation.systemsorting";
+
     static constexpr const char* sBatteryHidden              = "emulationstation.battery.hidden";
 
     static constexpr const char* sEsVideoMode                = "system.es.videomode";
@@ -437,4 +441,6 @@ class RecalboxConf: public IniFile, public StaticLifeCycleControler<RecalboxConf
     static const std::string& MenuFromEnum(Menu menu);
     static Relay RelayFromString(const std::string& relay);
     static const std::string& RelayFromEnum(Relay relay);
+    static SystemSorting SystemSortingFromString(const std::string& systemSorting);
+    static const std::string& SystemSortingFromEnum(SystemSorting systemSorting);
 };