+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@ as release notes for end user on a Recalbox upgrade.
 
 - Add delete game options
 - Add core libretro vitaquake2 for x64
+- New system sorting options in UI menu: sort your systems by type, manufacturer, release dates and more !
 
 ### Improvements